Add files via upload
A dynamic, reusable and shareable python code that a sports scientist who is working on Active Lives Survey Data and Rugby Football Union Data can use for data cleaning, data wrangling, data filtering, data preprocessing and data visualization. The code comes with a word file which consists of the instructions to use the said code.

The code is still a work in progress. So an updated instructions file along with a readme file will be uploaded soon.